+++ tests/integration/orm/has-many/instantiating-test.js
@@ -27,19 +27,19 @@ module('Integration | ORM | hasMany instantiating with params', {
       homeAddress: HomeAddress
     });
 
-    child1 = schema.homeAddress.find(1);
-    child2 = schema.homeAddress.find(2);
+    child1 = schema.homeAddresses.find(1);
+    child2 = schema.homeAddresses.find(2);
   }
 });
 
 test('children have fks added to their attrs', function(assert) {
-  let newChild = schema.homeAddress.new();
+  let newChild = schema.homeAddresses.new();
   assert.deepEqual(newChild.attrs, { userId: null });
   assert.deepEqual(child1.attrs, { id: '1', name: '123 Hyrule Way', userId: null });
 });
 
 test('the parent accepts an array of saved children ids', function(assert) {
-  let user = schema.user.new({ homeAddressIds: [1, 2] });
+  let user = schema.users.new({ homeAddressIds: [1, 2] });
 
   assert.equal(user.homeAddresses.models.length, 2);
   assert.deepEqual(user.homeAddresses.models[0], child1);
@@ -49,18 +49,18 @@ test('the parent accepts an array of saved children ids', function(assert) {
 
 test('the parent errors if one of the child ids doesnt exist', function(assert) {
   assert.throws(function() {
-    schema.user.new({ homeAddressIds: [1, 9] });
+    schema.users.new({ homeAddressIds: [1, 9] });
   }, /Couldn't find/);
 });
 
 test('the parent accepts an empty childIds array', function(assert) {
-  let user = schema.user.new({ homeAddressIds: [] });
+  let user = schema.users.new({ homeAddressIds: [] });
 
   assert.equal(user.homeAddresses.models.length, 0);
 });
 
 test('the parent accepts an array of saved child models', function(assert) {
-  let user = schema.user.new({ homeAddresses: [child1, child2] });
+  let user = schema.users.new({ homeAddresses: [child1, child2] });
 
   assert.deepEqual(user.homeAddressIds, ['1', '2']);
   assert.equal(user.homeAddresses.models.length, 2);
@@ -68,9 +68,9 @@ test('the parent accepts an array of saved child models', function(assert) {
 });
 
 test('the parent accepts an array of new child models', function(assert) {
-  let newAddress1 = schema.homeAddress.new();
-  let newAddress2 = schema.homeAddress.new();
-  let user = schema.user.new({ homeAddresses: [newAddress1, newAddress2] });
+  let newAddress1 = schema.homeAddresses.new();
+  let newAddress2 = schema.homeAddresses.new();
+  let user = schema.users.new({ homeAddresses: [newAddress1, newAddress2] });
 
   assert.deepEqual(user.homeAddressIds, [undefined, undefined]);
   assert.equal(user.homeAddresses.models.length, 2);
@@ -78,8 +78,8 @@ test('the parent accepts an array of new child models', function(assert) {
 });
 
 test('the parent accepts a mixed array of new and saved child models', function(assert) {
-  let newAddress1 = schema.homeAddress.new();
-  let user = schema.user.new({ homeAddresses: [child1, newAddress1] });
+  let newAddress1 = schema.homeAddresses.new();
+  let user = schema.users.new({ homeAddresses: [child1, newAddress1] });
 
   assert.deepEqual(user.homeAddressIds, ['1', undefined]);
   assert.equal(user.homeAddresses.models.length, 2);
@@ -88,21 +88,21 @@ test('the parent accepts a mixed array of new and saved child models', function(
 });
 
 test('the parent accepts null child models', function(assert) {
-  let user = schema.user.new({ addresses: [null] });
+  let user = schema.users.new({ addresses: [null] });
 
   assert.deepEqual(user.homeAddressIds, []);
   assert.equal(user.homeAddresses.models.length, 0);
 });
 
 test('the parent accepts no reference to a child id or model as empty obj', function(assert) {
-  let user = schema.user.new({});
+  let user = schema.users.new({});
 
   assert.deepEqual(user.homeAddressIds, []);
   assert.equal(user.homeAddresses.models.length, 0);
 });
 
 test('the parent accepts no reference to a child id or model', function(assert) {
-  let user = schema.user.new();
+  let user = schema.users.new();
 
   assert.deepEqual(user.homeAddressIds, []);
   assert.equal(user.homeAddresses.models.length, 0);