+++ tests/integration/orm/reinitialize-associations-test.js
@@ -18,16 +18,16 @@ module('Integration | ORM | reinitialize associations', {
       user: User
     });
 
-    this.schema.user.create({ id: 1, name: 'Link' });
-    this.schema.address.create({ id: 1, country: 'Hyrule', userId: 1 });
+    this.schema.users.create({ id: 1, name: 'Link' });
+    this.schema.addresses.create({ id: 1, country: 'Hyrule', userId: 1 });
   }
 });
 
 // By running two tests, we force the statically-defined classes to be
 // registered twice.
 test('safely initializes associations', function(assert) {
-  assert.equal(this.schema.user.find(1).addresses.models[0].country, 'Hyrule');
+  assert.equal(this.schema.users.find(1).addresses.models[0].country, 'Hyrule');
 });
 test('safely initializes associations again', function(assert) {
-  assert.equal(this.schema.user.find(1).addresses.models[0].country, 'Hyrule');
+  assert.equal(this.schema.users.find(1).addresses.models[0].country, 'Hyrule');
 });